Refactored create_game function into a Game class
