+++ src/components/Error.js
@@ -1,10 +1,11 @@
 import React from 'react';
+import Container from './Container';
 
 export default function Error() {
     return (
-        <div className={`overlay_content animate__animated animate__fadeIn animate__faster`}>
+        <Container animation='animate__fadeIn'>
             <h1 className="text_center">The server is not responding.</h1>
             <p className="text_center">Come back later and try again.</p>
-        </div>
+        </Container>
     );
 }
